+++ src/jp_learning_platform/application/__init__.py
@@ -1,0 +1,3 @@
+"""Application use-case layer."""
+
+__all__: list[str] = []